Various bugfixes and improvements
Now the progress shouldn't stop anywhere and should reload the page when finished

+++ generator/js/src/freenet/client/connection/SharedConnectionManager.java
@@ -133,8 +133,8 @@ public class SharedConnectionManager implements IConnectionManager, IUpdateManag
 		}
 	}
 	
-	private int getTime(){
-		return (int)(System.currentTimeMillis());
+	private long getTime(){
+		return System.currentTimeMillis();
 	}
 
 }
